Combine actions and util - search_index
